+++ rumqttd/src/server/broker.rs
@@ -1,18 +1,19 @@
 use crate::link::alerts::{self};
 use crate::link::console::ConsoleLink;
 use crate::link::network::{Network, N};
-use crate::link::remote::{self, RemoteLink};
+use crate::link::remote::{self, mqtt_connect, RemoteLink};
 use crate::link::{bridge, timer};
 use crate::local::LinkBuilder;
 use crate::protocol::v4::V4;
 use crate::protocol::v5::V5;
-use crate::protocol::Protocol;
+use crate::protocol::{Packet, Protocol};
 #[cfg(any(feature = "use-rustls", feature = "use-native-tls"))]
 use crate::server::tls::{self, TLSAcceptor};
 use crate::{meters, ConnectionSettings, Meter};
 use flume::{RecvError, SendError, Sender};
+use std::collections::HashMap;
 use std::net::{IpAddr, Ipv4Addr, SocketAddr};
-use std::sync::Arc;
+use std::sync::{Arc, Mutex};
 use tracing::{error, field, info, Instrument};
 
 #[cfg(feature = "websocket")]
@@ -33,7 +34,7 @@ use std::{io, thread};
 
 use crate::link::console;
 use crate::link::local::{self, LinkRx, LinkTx};
-use crate::router::{Disconnection, Event, Router};
+use crate::router::{Event, Router};
 use crate::{Config, ConnectionId, ServerSettings};
 use tokio::net::{TcpListener, TcpStream};
 use tokio::time::error::Elapsed;
@@ -184,7 +185,7 @@ impl Broker {
         // Spawn servers in a separate thread.
         for (_, config) in self.config.v4.clone() {
             let server_thread = thread::Builder::new().name(config.name.clone());
-            let server = Server::new(config, self.router_tx.clone(), V4);
+            let mut server = Server::new(config, self.router_tx.clone(), V4);
             server_thread.spawn(move || {
                 let mut runtime = tokio::runtime::Builder::new_current_thread();
                 let runtime = runtime.enable_all().build().unwrap();
@@ -200,7 +201,7 @@ impl Broker {
         if let Some(v5_config) = &self.config.v5 {
             for (_, config) in v5_config.clone() {
                 let server_thread = thread::Builder::new().name(config.name.clone());
-                let server = Server::new(config, self.router_tx.clone(), V5);
+                let mut server = Server::new(config, self.router_tx.clone(), V5);
                 server_thread.spawn(move || {
                     let mut runtime = tokio::runtime::Builder::new_current_thread();
                     let runtime = runtime.enable_all().build().unwrap();
@@ -219,7 +220,7 @@ impl Broker {
             for (_, config) in ws_config.clone() {
                 let server_thread = thread::Builder::new().name(config.name.clone());
                 //TODO: Add support for V5 procotol with websockets. Registered in config or on ServerSettings
-                let server = Server::new(config, self.router_tx.clone(), V4);
+                let mut server = Server::new(config, self.router_tx.clone(), V4);
                 server_thread.spawn(move || {
                     let mut runtime = tokio::runtime::Builder::new_current_thread();
                     let runtime = runtime.enable_all().build().unwrap();
@@ -294,10 +295,17 @@ pub enum LinkType {
     Remote,
 }
 
+#[derive(PartialEq)]
+enum AwaitingWill {
+    Cancel,
+    Fire,
+}
+
 struct Server<P> {
     config: ServerSettings,
     router_tx: Sender<(ConnectionId, Event)>,
     protocol: P,
+    awaiting_will_handler: Arc<Mutex<HashMap<String, Sender<AwaitingWill>>>>,
 }
 
 impl<P: Protocol + Clone + Send + 'static> Server<P> {
@@ -310,6 +318,7 @@ impl<P: Protocol + Clone + Send + 'static> Server<P> {
             config,
             router_tx,
             protocol,
+            awaiting_will_handler: Arc::new(Mutex::new(HashMap::default())),
         }
     }
 
@@ -327,7 +336,7 @@ impl<P: Protocol + Clone + Send + 'static> Server<P> {
         Ok((Box::new(stream), None))
     }
 
-    async fn start(&self, link_type: LinkType) -> Result<(), Error> {
+    async fn start(&mut self, link_type: LinkType) -> Result<(), Error> {
         let listener = TcpListener::bind(&self.config.listen).await?;
         let delay = Duration::from_millis(self.config.next_connection_delay_ms);
         let mut count: usize = 0;
@@ -376,24 +385,36 @@ impl<P: Protocol + Clone + Send + 'static> Server<P> {
                         }
                     };
                     task::spawn(
-                        remote(config, tenant_id.clone(), router_tx, stream, protocol).instrument(
-                            tracing::info_span!(
-                                "websocket_link",
-                                client_id = field::Empty,
-                                connection_id = field::Empty
-                            ),
-                        ),
+                        remote(
+                            config,
+                            tenant_id.clone(),
+                            router_tx,
+                            stream,
+                            protocol,
+                            self.awaiting_will_handler.clone(),
+                        )
+                        .instrument(tracing::info_span!(
+                            "websocket_link",
+                            client_id = field::Empty,
+                            connection_id = field::Empty
+                        )),
                     )
                 }
                 LinkType::Remote => task::spawn(
-                    remote(config, tenant_id.clone(), router_tx, network, protocol).instrument(
-                        tracing::error_span!(
-                            "remote_link",
-                            ?tenant_id,
-                            client_id = field::Empty,
-                            connection_id = field::Empty,
-                        ),
-                    ),
+                    remote(
+                        config,
+                        tenant_id.clone(),
+                        router_tx,
+                        network,
+                        protocol,
+                        self.awaiting_will_handler.clone(),
+                    )
+                    .instrument(tracing::error_span!(
+                        "remote_link",
+                        ?tenant_id,
+                        client_id = field::Empty,
+                        connection_id = field::Empty,
+                    )),
                 ),
             };
 
@@ -431,26 +452,74 @@ async fn remote<P: Protocol>(
     router_tx: Sender<(ConnectionId, Event)>,
     stream: Box<dyn N>,
     protocol: P,
+    will_handlers: Arc<Mutex<HashMap<String, Sender<AwaitingWill>>>>,
 ) {
-    let network = Network::new(
+    let mut network = Network::new(
         stream,
         config.max_payload_size,
         config.max_inflight_count,
         protocol,
     );
-    // Start the link
-    let mut link =
-        match RemoteLink::new(config, router_tx.clone(), tenant_id.clone(), network).await {
-            Ok(l) => l,
-            Err(e) => {
-                error!(error=?e, "Remote link error");
-                return;
-            }
-        };
 
-    let client_id = link.client_id.to_owned();
+    let dynamic_filters = config.dynamic_filters;
+
+    let connect_packet = match mqtt_connect(config, &mut network).await {
+        Ok(p) => p,
+        Err(e) => {
+            error!(error=?e, "Error while handling MQTT connect packet");
+            return;
+        }
+    };
+
+    let (mut client_id, clean_session) = match &connect_packet {
+        Packet::Connect(ref connect, _, _, _, _) => {
+            (connect.client_id.clone(), connect.clean_session)
+        }
+        _ => unreachable!(),
+    };
+
+    if let Some(tenant_id) = &tenant_id {
+        // client_id is set to "tenant_id.client_id"
+        // this is to make sure we are consistent,
+        // as Connection uses this format of client_id
+        client_id = format!("{tenant_id}.{client_id}");
+    }
+
+    if let Some(sender) = will_handlers.lock().unwrap().remove(&client_id) {
+        let awaiting_will = if clean_session {
+            AwaitingWill::Fire
+        } else {
+            AwaitingWill::Cancel
+        };
+        sender.try_send(awaiting_will).unwrap();
+    }
+
+    let (will_tx, will_rx) = flume::bounded::<AwaitingWill>(1);
+    will_handlers
+        .lock()
+        .unwrap()
+        .insert(client_id.clone(), will_tx);
+
+    // Start the link
+    let mut link = match RemoteLink::new(
+        router_tx.clone(),
+        tenant_id.clone(),
+        network,
+        connect_packet,
+        dynamic_filters,
+    )
+    .await
+    {
+        Ok(l) => l,
+        Err(e) => {
+            error!(error=?e, "Remote link error");
+            return;
+        }
+    };
+
     let connection_id = link.connection_id;
-    let mut execute_will = false;
+    let will_delay_interval = link.will_delay_interval;
+    let mut send_disconnect = true;
 
     match link.start().await {
         // Connection got closed. This shouldn't usually happen.
@@ -459,22 +528,45 @@ async fn remote<P: Protocol>(
         // originated internally in the router.
         Err(remote::Error::Link(e)) => {
             error!(error=?e, "router-drop");
-            return;
+            send_disconnect = false;
         }
         // Any other error
         Err(e) => {
             error!(error=?e, "Disconnected!!");
-            execute_will = true;
         }
     };
 
-    let disconnect = Disconnection {
-        id: client_id,
-        execute_will,
-        pending: vec![],
+    if send_disconnect {
+        let disconnect = Event::Disconnect;
+        let message = (connection_id, disconnect);
+        router_tx.send(message).ok();
+    }
+
+    // this is important to stop the connection
+    drop(link);
+
+    let publish_will = match tokio::time::timeout(
+        Duration::from_secs(will_delay_interval as u64),
+        will_rx.recv_async(),
+    )
+    .await
+    {
+        Ok(w) => w.is_ok_and(|k| k == AwaitingWill::Fire),
+        Err(_) => {
+            // no need to keep the sender after timeout
+            will_handlers.lock().unwrap().remove(&client_id);
+            // as will delay interval has passed, publish the will message
+            true
+        }
     };
 
-    let disconnect = Event::Disconnect(disconnect);
-    let message = (connection_id, disconnect);
-    router_tx.send(message).ok();
+    if publish_will {
+        let message = Event::PublishWill((client_id, tenant_id));
+        // is this connection_id really correct at this point?
+        // as we have disconnected already, some other connection
+        // might be using this connection ID!
+        // It won't matter in this case as we don't use it
+        // but might affect logs?
+        router_tx.send((connection_id, message)).ok();
+    }
 }