+++ rumqttd/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Session expiry and clean start for MQTTv5 (#691)
+- Will delay interval for MQTTv5 (#686)
 
 ### Changed
 
@@ -18,7 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Link and its implementation which were deprecated.
 
 ### Fixed
+- Will Messages
 - Retained Messages
+- Publish properties in QoS2 publish
 
 ### Security